pyproject.toml
[tool.poetry]
name = "psycopg2-2-9-create-database"
version = "0.1.0"
description = ""
authors = ["sinofseven <em.s.00001@gmail.com>"]
readme = "README.md"
packages = [{include = "psycopg2_2"}]

[tool.poetry.dependencies]
python = "^3.8"
psycopg2-binary = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
